requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
Pillow==5.1.0
requests==2.19.1
tinify==1.5.1
urllib3==1.23
